@@ -1,16 +1,22 @@
 package ru.javawebinar.topjava.model;
 
+import org.hibernate.Hibernate;
+import org.springframework.data.domain.Persistable;
+
 import javax.persistence.*;
 
-    @MappedSuperclass
-    // http://stackoverflow.com/questions/594597/hibernate-annotations-which-is-better-field-or-property-access
-    @Access(AccessType.FIELD)
-    public abstract class AbstractBaseEntity {
+@MappedSuperclass
+// http://stackoverflow.com/questions/594597/hibernate-annotations-which-is-better-field-or-property-access
+@Access(AccessType.FIELD)
+public abstract class AbstractBaseEntity implements Persistable<Integer> {
     public static final int START_SEQ = 100000;
 
     @Id
     @SequenceGenerator(name = "global_seq", sequenceName = "global_seq", allocationSize = 1, initialValue = START_SEQ)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "global_seq")
+
+//  See https://hibernate.atlassian.net/browse/HHH-3718 and https://hibernate.atlassian.net/browse/HHH-12034
+//  Proxy initialization when accessing its identifier managed now by JPA_PROXY_COMPLIANCE setting
     protected Integer id;
 
     protected AbstractBaseEntity() {
@@ -20,14 +26,16 @@ import javax.persistence.*;
         this.id = id;
     }
 
-    public Integer getId() {
-        return id;
-    }
-
     public void setId(Integer id) {
         this.id = id;
     }
 
+    @Override
+    public Integer getId() {
+        return id;
+    }
+
+    @Override
     public boolean isNew() {
         return this.id == null;
     }
@@ -43,7 +51,7 @@ import javax.persistence.*;
         if (this == o) {
             return true;
         }
-        if (o == null || getClass() != o.getClass()) {
+        if (o == null || !getClass().equals(Hibernate.getClass(o))) {
             return false;
         }
         AbstractBaseEntity that = (AbstractBaseEntity) o;
